requirements.txt
matplotlib==3.3.4
nltk==3.6.1
numpy==1.19.5
pandas==1.2.4
scikit_learn==1.0.2
streamlit==1.4.0
tensorflow-cpu==2.7.0
